improve auto scroll on dnd
