+++ web/src/beta/components/DragAndDropList/index.tsx
@@ -1,9 +1,11 @@
 import type { ReactNode } from "react";
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
+import { useDragDropManager } from "react-dnd";
 
 import { styled } from "@reearth/services/theme";
 
 import Item from "./Item";
+import { useScroll } from "./scrollItem";
 
 type Props<Item extends { id: string } = { id: string }> = {
   uniqueKey: string;
@@ -23,6 +25,20 @@ function DragAndDropList<Item extends { id: string } = { id: string }>({
   gap,
 }: Props<Item>) {
   const [movingItems, setMovingItems] = useState<Item[]>(items);
+  const listRef = useRef<null | HTMLDivElement>(null);
+
+  const { updatePosition } = useScroll(listRef);
+
+  const dragDropManager = useDragDropManager();
+  const monitor = dragDropManager.getMonitor();
+
+  useEffect(() => {
+    const unsubscribe = monitor.subscribeToOffsetChange(() => {
+      const offset = monitor.getSourceClientOffset()?.y as number;
+      updatePosition({ position: offset, isScrollAllowed: true });
+    });
+    return unsubscribe;
+  }, [monitor, updatePosition]);
 
   const customDragHandler = (item: Item): boolean => {
     // eslint-disable-next-line no-prototype-builtins
@@ -55,7 +71,7 @@ function DragAndDropList<Item extends { id: string } = { id: string }>({
   }, [items]);
 
   return (
-    <SWrapper gap={gap}>
+    <SWrapper gap={gap} ref={listRef}>
       {movingItems.map((item, i) => {
         const id = getId(item);
         const shouldUseCustomHandler = customDragHandler(item); // Determine whether to use custom handler
